+++ pkg/controller/servicebindingrequest/retriever.go
@@ -6,7 +6,6 @@ import (
 	"strings"
 
 	"github.com/go-logr/logr"
-	"github.com/redhat-developer/service-binding-operator/pkg/controller/servicebindingrequest/planner"
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -19,7 +18,7 @@ import (
 type Retriever struct {
 	ctx           context.Context   // request context
 	client        client.Client     // Kubernetes API client
-	plan          *planner.Plan     // plan instance
+	plan          *Plan             // plan instance
 	logger        logr.Logger       // logger instance
 	data          map[string][]byte // data retrieved
 	volumeKeys    []string
@@ -230,7 +229,7 @@ func (r *Retriever) Retrieve() error {
 }
 
 // NewRetriever instantiate a new retriever instance.
-func NewRetriever(ctx context.Context, client client.Client, plan *planner.Plan, bindingPrefix string) *Retriever {
+func NewRetriever(ctx context.Context, client client.Client, plan *Plan, bindingPrefix string) *Retriever {
 	return &Retriever{
 		ctx:           ctx,
 		client:        client,
